Clean up classes and make packages more sensible
